Update paths for ref
