+++ gwemopt/data/config/ZTF.config
@@ -10,7 +10,7 @@ FOV_coverage_type square
 FOV_type square
 tesselationFile ZTF.tess
 quadrantsFile ZTF.quadrants
-referenceFile ../input/ZTF.ref
+referenceFile ZTF.ref
 slew_rate 2.6
 readout 10
 horizon 30
