+++ gwemopt/paths.py
@@ -11,5 +11,6 @@ DEFAULT_LIGHTCURVE_DIR = DATA_DIR.joinpath("lightcurves")
 DEFAULT_BASE_OUTPUT_DIR = Path.home().joinpath("Data/gwemopt/")
 
 TESSELATION_DIR = DATA_DIR.joinpath("tesselations")
+REFS_DIR = DATA_DIR.joinpath("refs")
 
 test_skymap = DATA_DIR.joinpath("skymaps/S190425z_2_LALInference.fits.gz")
